Add project description
About this project
I will preface this by saying it is my first project, and so it will most likely be improved upon in the future.

When it comes to this Sales Analysis, the purpose was to analyze a sample of Amazon order data (which can be found here: https://www.kaggle.com/datasets/knightbearr/sales-product-data) and try and gather KPI's to better understand Amazon's customers and their buying patterns/preferences. This could in turn allow Amazon to make more informed decisions in regards to what products to sell and when to advertise their products (as examples).

In order to help Amazon better understand their customers, some key questions needed to be answered to help identify trends in the data.

Key questions include: What is the total number of sales? What is the average sales per month? What is the monthly revenue? Which market (state) generated the most sales on average? When were the best- and worst-selling periods (quarters)? What City had the highest number of sales? Which products sell best? What products are most often sold together? Which products should the company order more or less of? What times should we display advertisement to maximize likelihood of customer's buying product?

From answering these questions, several key insights were found:

    Most commonly bought products pairs by far were the iPhone and Lightning Charging Cable (1014 times) and the Google Phone with USB-C Charging Cable (999 times). Both pairs contain a main product (the phone) and a compliment product (charging cable).
    California generated the most sales by far (having over 3x the sales of the next closest state, New York).
    San Francisco had the most sales, followed by Los Angeles (both California cities).
    June to September was the weakest period for sales, while October to December was the strongest period.
    Macbook Pro Laptops had 2x the sales of the next closest product (iPhone).
    Time ranges of 11am to 1pm, and 6pm to 8pm are the best times for advertisements (7pm is when the most products are bought).

Based on this analysis, Amazon now can better understand their customers (including product preferences, where they commonly buy from, and up and downtimes for customer sales) and come up with solutions to address potential problem areas (such as the slump in sales during the summer).

An example of a potential solution could be Amazon hosting their Prime Day during the summer months to try and boost sales which they have done over the previous years. This same approach could be used from January to March (sales are low during these months as well) through offering various discounts on in-demand items.

Hopefully you enjoy looking through this project, and I'm always open to feedback.